requirements/base.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
loguru>=0.7.0
PyYAML>=6.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
pyyaml>=6.0
